Feat: Soporte ventas con cantidades decimales + Mejoras ticket
VENTAS CON DECIMALES:
- Campo qty en salesItems: IntegerField → DecimalField(10,3)
- Migración aplicada en pos
- Acepta coma y punto en inputs de cantidad (1,5 o 1.5)
- Conversión automática coma → punto en JavaScript
- Métodos update_quantity_on_sale usan Decimal

POS - INPUTS:
- Campos cantidad cambiados de type=number a type=text
- Aceptan formato argentino con coma (1,25)
- Conversión en calc() y agregar item

TICKET TÉRMICO 58MM:
- Color negro sólido (#000) en lugar de gris
- Forzar impresión en negro (-webkit-print-color-adjust)
- Cantidad muestra 2 decimales (1.50 kg)
- Mejora legibilidad en impresoras térmicas

EDICIÓN RÁPIDA PRECIOS:
- Filtro automático al escribir en buscador
- Filtro automático al cambiar proveedor
- Filtro automático al cambiar orden
- Sin necesidad de botón Aplicar (pero sigue funcionando)

COMPATIBILIDAD:
- Todos los reportes compatibles con qty decimal
- Cálculos de ganancia usan Decimal(item.qty)
- Sin impacto en reportes PDF/Excel existentes

+++ store/pos/models.py
@@ -80,7 +80,7 @@ class salesItems(models.Model):
     product = models.ForeignKey(Products, on_delete=models.CASCADE)
     price = models.FloatField(default=0)  # Precio al momento de la venta
     costo_unitario = models.FloatField(default=0)  # Costo al momento de la venta
-    qty = models.IntegerField(default=0)
+    qty = models.DecimalField(max_digits=10, decimal_places=3, default=0)
     total = models.FloatField(default=0)
 
     def save(self, *args, **kwargs):
